test: add MainViewModel navigation coverage
