+++ ToolManagementAppV2.Tests/ViewModels/ToolManagementViewModelTests.cs
@@ -1,0 +1,97 @@
+using System.IO;
+using System.Linq;
+using ToolManagementAppV2.Models.Domain;
+using ToolManagementAppV2.Services.Core;
+using ToolManagementAppV2.Services.Tools;
+using ToolManagementAppV2.Interfaces;
+using ToolManagementAppV2.ViewModels;
+using Xunit;
+
+namespace ToolManagementAppV2.Tests.ViewModels
+{
+    public class ToolManagementViewModelTests
+    {
+        [Fact]
+        public void SearchCommand_FiltersToolsBySearchTerm()
+        {
+            var dbPath = Path.GetTempFileName();
+            try
+            {
+                var db = new DatabaseService(dbPath);
+                IToolService toolService = new ToolService(db);
+                var vm = new ToolManagementViewModel(toolService);
+                toolService.AddTool(new Tool { ToolNumber = "T1", NameDescription = "Hammer" });
+                toolService.AddTool(new Tool { ToolNumber = "T2", NameDescription = "Saw" });
+                vm.SearchTerm = "Ham";
+                vm.SearchCommand.Execute(null);
+                Assert.Single(vm.SearchResults);
+                Assert.Equal("Hammer", vm.SearchResults.First().NameDescription);
+            }
+            finally
+            {
+                if (File.Exists(dbPath))
+                    File.Delete(dbPath);
+            }
+        }
+
+        [Fact]
+        public void SearchCommand_SortsResultsIntoCategories()
+        {
+            var dbPath = Path.GetTempFileName();
+            try
+            {
+                var db = new DatabaseService(dbPath);
+                IToolService toolService = new ToolService(db);
+                var vm = new ToolManagementViewModel(toolService);
+                toolService.AddTool(new Tool { ToolNumber = "T1", NameDescription = "Hammer" });
+                toolService.AddTool(new Tool { ToolNumber = "T2", NameDescription = "Cordless Drill" });
+                vm.SearchTerm = string.Empty;
+                vm.SearchCommand.Execute(null);
+                Assert.Single(vm.HandTools);
+                Assert.Single(vm.PowerTools);
+            }
+            finally
+            {
+                if (File.Exists(dbPath))
+                    File.Delete(dbPath);
+            }
+        }
+
+        [Fact]
+        public void AddToolCommand_PersistsNewToolValues()
+        {
+            var dbPath = Path.GetTempFileName();
+            try
+            {
+                var db = new DatabaseService(dbPath);
+                IToolService toolService = new ToolService(db);
+                var vm = new ToolManagementViewModel(toolService);
+                vm.NewTool.ToolNumber = "TN1";
+                vm.NewTool.NameDescription = "Hammer";
+                vm.NewTool.PartNumber = "PN1";
+                vm.NewTool.Brand = "BrandA";
+                vm.NewTool.Location = "Shelf";
+                vm.NewTool.QuantityOnHand = 5;
+                vm.NewTool.Supplier = "ABC";
+                vm.NewTool.Notes = "Note";
+                vm.AddToolCommand.Execute(null);
+                var tools = toolService.GetAllTools();
+                Assert.Single(tools);
+                var tool = tools.First();
+                Assert.Equal("TN1", tool.ToolNumber);
+                Assert.Equal("Hammer", tool.NameDescription);
+                Assert.Equal("PN1", tool.PartNumber);
+                Assert.Equal("BrandA", tool.Brand);
+                Assert.Equal("Shelf", tool.Location);
+                Assert.Equal(5, tool.QuantityOnHand);
+                Assert.Equal("ABC", tool.Supplier);
+                Assert.Equal("Note", tool.Notes);
+            }
+            finally
+            {
+                if (File.Exists(dbPath))
+                    File.Delete(dbPath);
+            }
+        }
+    }
+}
